refactor(backend): make normal web apis use new auth module
