+++ apps/backend/src/auth/webapis/list-sessions.ts
@@ -1,5 +1,6 @@
-import { sessionDigestSchema, SessionService } from '~backend/services/auth/session';
 import { Static, t } from 'elysia';
+import { sessionDigestSchema } from '~backend/auth/auth.entities';
+import { SessionService } from '~backend/services/auth/session';
 
 export const listSessionsResponseSchema = t.Object({
 	sessions: t.Array(sessionDigestSchema),
@@ -9,6 +10,6 @@ export async function listSessions(
 	userId: string,
 ): Promise<Static<typeof listSessionsResponseSchema>> {
 	return {
-		sessions: await SessionService.getUserSessions(userId),
+		sessions: await SessionService.findUserSessions(userId),
 	};
 }
