+++ apps/backend/src/auth/webapis/refresh.ts
@@ -1,21 +1,16 @@
-import {
-	Session,
-	sessionSchema,
-	SessionScope,
-	SessionService,
-} from '~backend/services/auth/session';
 import { Static, t } from 'elysia';
+import { Session, sessionSchema, SessionScope } from '~backend/auth/auth.entities';
+import { SessionService } from '~backend/services/auth/session';
 
 export const refreshResponseSchema = t.Object({
 	session: sessionSchema,
 });
 
 export async function refresh(oldSession: Session): Promise<Static<typeof refreshResponseSchema>> {
-	await SessionService.invalidate(oldSession.id);
+	await SessionService.revoke(oldSession.id);
 
 	const session = await SessionService.create(oldSession.userId, {
 		scope: SessionScope.DEFAULT,
-		metadata: {},
 	});
 
 	return {
