+++ apps/backend/src/files/files.routes.ts
@@ -1,7 +1,7 @@
 import { Elysia } from 'elysia';
 import { authMiddleware } from '~backend/shared/auth/middleware';
-import { SessionScope } from '~backend/services/auth/session';
 import { upload, uploadBodySchema, uploadResponseSchema } from './webapis/upload';
+import { SessionScope } from '~backend/auth/auth.entities';
 
 export const FilesRoutes = new Elysia({
 	name: 'Routes.Files',
