+++ apps/backend/src/textures/textures.routes.ts
@@ -1,9 +1,9 @@
 import { Elysia } from 'elysia';
 import { authMiddleware } from '~backend/shared/auth/middleware';
-import { SessionScope } from '~backend/services/auth/session';
 import { create, createBodySchema, createResponseSchema } from './webapis/create';
 import { inspect, inspectParamsSchema, inspectResponseSchema } from './webapis/inspect';
 import { getImage, getImageParamsSchema, getImageResponseSchema } from './webapis/get-image';
+import { SessionScope } from '~backend/auth/auth.entities';
 
 export const TexturesRoutes = new Elysia({ name: 'Routes.Textures', prefix: '/textures' })
 	.get('/:id', async ({ params }) => await inspect(params), {
